Removed compiler warnings. State machine binary tree now linked to layer 2 object (layer 2 and tei). They will now be created dynamically for each layer 2 object.
Modified Files:
Tag: mISDNuser_1_0
	mISDNuser/i4lnet/net_l2.c mISDNuser/i4lnet/net_l2.h
 	mISDNuser/i4lnet/net_l3.h mISDNuser/i4lnet/tei.c
	mISDNuser/include/isdn_net.h mISDNuser/voip/voip_appl.c

+++ i4lnet/net_l3.h
@@ -30,7 +30,7 @@ struct _layer3_proc {
 	int		ces;
 	int		selces;
 	int		state;
-	int		Flags;
+	u_long		Flags;
 	L3Timer_t	timer1;
 	L3Timer_t	timer2;
 	int		bc;
